+++ src/main/java/com/b2b/orderandinventory/exception/OrderDoesNotExistException.java
@@ -1,9 +1,7 @@
 package com.b2b.orderandinventory.exception;
 
-import java.math.BigInteger;
-
 public class OrderDoesNotExistException extends RuntimeException {
-    public OrderDoesNotExistException(BigInteger orderId) {
+    public OrderDoesNotExistException(Long orderId) {
         super("No Order with id: "+orderId);
     }
 }
