+++ src/main/java/com/b2b/orderandinventory/model/Company.java
@@ -1,19 +1,24 @@
 package com.b2b.orderandinventory.model;
 
-import jakarta.persistence.*;
+import java.time.LocalDateTime;
+
 import org.springframework.data.annotation.CreatedDate;
 import org.springframework.data.annotation.LastModifiedDate;
 import org.springframework.data.jpa.domain.support.AuditingEntityListener;
 
-import java.math.BigInteger;
-import java.time.LocalDateTime;
+import jakarta.persistence.Column;
+import jakarta.persistence.Entity;
+import jakarta.persistence.EntityListeners;
+import jakarta.persistence.GeneratedValue;
+import jakarta.persistence.GenerationType;
+import jakarta.persistence.Id;
 
 @Entity
 @EntityListeners(AuditingEntityListener.class)
 public class Company {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private BigInteger id;
+    private Long id;
 
     @Column(nullable = false)
     private String name;
@@ -32,7 +37,7 @@ public class Company {
     public Company() {
     }
 
-    public Company(BigInteger id, String name, String address, LocalDateTime createAt, LocalDateTime lastUpdatedAt) {
+    public Company(Long id, String name, String address, LocalDateTime createAt, LocalDateTime lastUpdatedAt) {
         this.id = id;
         this.name = name;
         this.address = address;
@@ -40,11 +45,11 @@ public class Company {
         this.lastUpdatedAt = lastUpdatedAt;
     }
 
-    public BigInteger getId() {
+    public Long getId() {
         return id;
     }
 
-    public void setId(BigInteger id) {
+    public void setId(Long id) {
         this.id = id;
     }
